refactor: extract suppress_logging to shared logger_utils module
